+++ crates/uselesskey-core-rustls-pki/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
 name = "uselesskey-core-rustls-pki"
-version = "0.7.1"
+version = "0.7.2"
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 repository.workspace = true
-description = "rustls-pki-types adapters for uselesskey fixture crates."
+description = "Deprecated compatibility shim re-exporting rustls-pki extension traits from uselesskey-rustls."
 categories.workspace = true
 keywords = ["rustls", "pki", "fixtures", "testing", "adapter"]
 readme = "README.md"
@@ -15,30 +15,15 @@ documentation = "https://docs.rs/uselesskey-core-rustls-pki"
 authors.workspace = true
 
 [dependencies]
-rustls-pki-types = "1"
-
-# Key type crates (all optional)
-uselesskey-x509 = { path = "../uselesskey-x509", version = "0.7.1", optional = true }
-uselesskey-rsa = { path = "../uselesskey-rsa", version = "0.7.1", optional = true }
-uselesskey-ecdsa = { path = "../uselesskey-ecdsa", version = "0.7.1", optional = true }
-uselesskey-ed25519 = { path = "../uselesskey-ed25519", version = "0.7.1", optional = true }
+uselesskey-rustls.workspace = true
 
 [features]
 default = ["x509"]
-x509 = ["dep:uselesskey-x509"]
-rsa = ["dep:uselesskey-rsa"]
-ecdsa = ["dep:uselesskey-ecdsa"]
-ed25519 = ["dep:uselesskey-ed25519"]
+x509 = ["uselesskey-rustls/x509"]
+rsa = ["uselesskey-rustls/rsa"]
+ecdsa = ["uselesskey-rustls/ecdsa"]
+ed25519 = ["uselesskey-rustls/ed25519"]
 all = ["x509", "rsa", "ecdsa", "ed25519"]
-
-[dev-dependencies]
-insta.workspace = true
-serde = { workspace = true, features = ["derive"] }
-uselesskey-core = { path = "../uselesskey-core", version = "0.7.1" }
-uselesskey-x509 = { path = "../uselesskey-x509", version = "0.7.1" }
-uselesskey-rsa = { path = "../uselesskey-rsa", version = "0.7.1" }
-uselesskey-ecdsa = { path = "../uselesskey-ecdsa", version = "0.7.1" }
-uselesskey-ed25519 = { path = "../uselesskey-ed25519", version = "0.7.1" }
 
 [package.metadata.docs.rs]
 features = ["all"]
