+++ crates/uselesskey-core-rustls-pki/README.md
@@ -1,6 +1,11 @@
-# uselesskey-core-rustls-pki
+Deprecated compatibility shim.
 
-Core adapter traits that convert uselesskey fixtures into `rustls-pki-types` key and certificate types.
+The canonical rustls-pki conversion traits now live in `uselesskey-rustls`:
 
-This crate only covers PKI conversion primitives (`PrivateKeyDer`, `CertificateDer`).
-Higher-level rustls configuration builders live in `uselesskey-rustls`.
+```rust
+use uselesskey_rustls::{RustlsPrivateKeyExt, RustlsCertExt, RustlsChainExt};
+```
+
+Prefer `uselesskey-rustls` (or the `uselesskey` facade) for supported PKI
+fixture-conversion APIs. This crate is retained only to keep v0.7.x consumers
+compiling and is scheduled for removal in a later v0.8.x PR.
