+++ crates/uselesskey-rustls/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "uselesskey-rustls"
-version = "0.7.1"
+version = "0.7.2"
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
@@ -15,7 +15,7 @@ documentation = "https://docs.rs/uselesskey-rustls"
 authors.workspace = true
 
 [dependencies]
-uselesskey-core-rustls-pki = { path = "../uselesskey-core-rustls-pki", version = "0.7.1", default-features = false }
+rustls-pki-types = "1"
 rustls = { version = "0.23", default-features = false, features = ["std", "tls12"], optional = true }
 
 # Key type crates (all optional)
@@ -29,10 +29,10 @@ features = ["all", "tls-config", "rustls-ring"]
 
 [features]
 default = ["x509"]
-x509 = ["uselesskey-core-rustls-pki/x509", "dep:uselesskey-x509"]
-rsa = ["uselesskey-core-rustls-pki/rsa", "dep:uselesskey-rsa"]
-ecdsa = ["uselesskey-core-rustls-pki/ecdsa", "dep:uselesskey-ecdsa"]
-ed25519 = ["uselesskey-core-rustls-pki/ed25519", "dep:uselesskey-ed25519"]
+x509 = ["dep:uselesskey-x509"]
+rsa = ["dep:uselesskey-rsa"]
+ecdsa = ["dep:uselesskey-ecdsa"]
+ed25519 = ["dep:uselesskey-ed25519"]
 all = ["x509", "rsa", "ecdsa", "ed25519"]
 server-config = ["dep:rustls", "x509"]
 client-config = ["dep:rustls", "x509"]
