+++ crates/uselesskey-rustls/src/lib.rs
@@ -2,9 +2,11 @@
 
 //! Integration between uselesskey test fixtures and `rustls-pki-types`.
 //!
-//! This crate re-exports extension traits from `uselesskey-core-rustls-pki`
-//! that convert uselesskey fixtures into `rustls-pki-types` types
-//! (`PrivateKeyDer`, `CertificateDer`).
+//! This crate owns the PKI extension traits that convert uselesskey
+//! fixtures into `rustls-pki-types` types (`PrivateKeyDer`,
+//! `CertificateDer`). The implementation lives under
+//! [`crate::srp::pki`]; the published-internal `uselesskey-core-rustls-pki`
+//! crate re-exports the same surface for v0.7.x consumers.
 //!
 //! With the `server-config` and `client-config` features, it also provides
 //! convenience builders for `rustls::ServerConfig` and `rustls::ClientConfig`,
@@ -40,12 +42,15 @@
 #[cfg(any(feature = "server-config", feature = "client-config"))]
 mod config;
 
+#[doc(hidden)]
+pub mod srp;
+
 #[cfg(test)]
 mod testutil;
 
 #[cfg(feature = "x509")]
-pub use uselesskey_core_rustls_pki::RustlsChainExt;
-pub use uselesskey_core_rustls_pki::{RustlsCertExt, RustlsPrivateKeyExt};
+pub use srp::pki::RustlsChainExt;
+pub use srp::pki::{RustlsCertExt, RustlsPrivateKeyExt};
 
 #[cfg(feature = "server-config")]
 pub use config::RustlsServerConfigExt;
