+++ crates/uselesskey-rustls/src/srp/mod.rs
@@ -1,0 +1,6 @@
+//! Single-responsibility rustls adapter internals.
+//!
+//! These modules preserve the old published internal crate APIs while moving
+//! the implementation behind the `uselesskey-rustls` public package promise.
+
+pub mod pki;
