+++ docs/reference/support-matrix.md
@@ -31,7 +31,7 @@
 | `uselesskey-core-negative` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; generic negative helpers are owned by `uselesskey-core`. | Tracks workspace MSRV (Rust 1.95). | Prefer `uselesskey::negative` or `uselesskey-core`; this crate is retained only as a compatibility shim. |
 | `uselesskey-core-negative-der` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; DER corruption helpers are owned by `uselesskey-core`. | Tracks workspace MSRV (Rust 1.95). | Prefer `uselesskey::negative` or `uselesskey-core`; this crate is retained only as a compatibility shim. |
 | `uselesskey-core-negative-pem` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; PEM corruption helpers are owned by `uselesskey-core`. | Tracks workspace MSRV (Rust 1.95). | Prefer `uselesskey::negative` or `uselesskey-core`; this crate is retained only as a compatibility shim. |
-| `uselesskey-core-rustls-pki` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.95). | Prefer depending on `uselesskey` or key-type facade crates unless you are extending the workspace. |
+| `uselesskey-core-rustls-pki` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; the rustls-pki adapter traits are owned by `uselesskey-rustls`. | Tracks workspace MSRV (Rust 1.95). | Prefer `uselesskey-rustls`; this crate is retained only as a compatibility shim. |
 | `uselesskey-core-seed` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; seed parsing and redaction are owned by `uselesskey-core`. | Tracks workspace MSRV (Rust 1.95). | Prefer `uselesskey-core::Seed`; this crate is retained only as a compatibility shim. |
 | `uselesskey-core-sink` | `experimental` | `published` | — | `repo-internal` | Published-internal compatibility shim; tempfile sink mechanics are owned by `uselesskey-core`. | Tracks workspace MSRV (Rust 1.95). | Prefer `uselesskey::TempArtifact` or `uselesskey-core`; this crate is retained only as a compatibility shim. |
 | `uselesskey-core-token` | `experimental` | `published` | — | `repo-internal` | Internal building blocks; APIs may change between minor releases. | Tracks workspace MSRV (Rust 1.95). | Compatibility shim; supported token fixture APIs now live under `uselesskey-token`. |
